Add initial support for XML version of list of meetings.
This patch adds initial support for an XML version of the list of
meetings, in the form of a Python script meetings.py that can convert
the fixed-width text version into such an XML version, and can also
read in that XML version again.

Once meetings.py has been extended to cover all the functionality of
the old parselist script, the XML version is intended to be checked in
to replace the fixed-width text list as the master source for the list
of meetings.  After that, additional information about talks and
speakers can be added that does not fit into the fixed-width list.