Support listing speakers by number of talks in meetings.py.
Part of the process of replicating the functionality of the old
parselist script in terms of the XML meeting list.  As with the old
script, speaker names (without titles), as in the list, are treated as
the key for identifying speakers, so any speakers appearing under more
than one name will appear separately in these statistics.