Support converting XML list of meetings to legacy text form.
This is for one-off validation that the conversion process from text
to XML does not lose information.  It does not produce results
byte-identical to the original text file, or fully validly formatted
suitably for reading back in, just similar enough for comparison.
Specifically, there are differences in how some unminuted meetings are
reflected in the text file, and this output does not truncate too-long
meeting titles (so the output for those meetings is not actually valid
input for the text-to-XML conversion).