Add support for linking to extra speaker information.
The links go in a separate file speakers.xml, since they are
information about a speaker shared between all their talks.  Currently
the unique id of a speaker for this file is always determined as
"last, first", but in future it could be possible to list that
explicitly in the list of meetings for cases of speakers who changed
their names, or of more than one speaker with the same name.

In the initial version, only a small selection of speakers have links
present, and Wikipedia links are the only kind present.